filter out rows of matrix with only one observation
